Get rid of error message after COG jobs successfully complete.
Even when the integrated alerts COG jobs successfully complete, there is
a data-updates error message. It turns out this is because we didn't
increase the integrated_alerts job timeout when we added the very long
extra COG steps. So, I just added an extra 7 hours to the timeout for
IntegratedAlertsSync job.