Put in retries for ThrottlingException for describe_cluster() call
We seem to seem to get a ThrottlingException on describe_cluster()
fairly often (often happens several time every couple of nights). This
causes the entire job process to fail (or at least not finish the
post-processing). So, I'm adding a few retries if we get a ClientError
which is a ThrottlingException on the describe_cluster() call in
check_analysis().