Increase time for the unify_projection part of dist alert processing
The initial ingestion job for dist alerts (which includes
unify_projection) usually finishes in under 2 hours, but every few
times, it goes over. So increase the timeout to 3 hours.

This is a hot fix, since there are a bunch of other hotfixes as well.
Then I will merge all the hotfixes back to develop.