For user area sync, fetch areas sorted by newest first.
This is a merge from Justin's PR from develop to master, but I fixed
some things and added a bunch more since checking in Justin's original
PR, so making this a separate PR to be more explicit and so we can
modify it further as needed, like a hotfix.

This started with Justin's original PR to give start/end parts to the
sync command. Date ranges on the sync command turned out to not be that
important, but I left it in, in case we do want to specify explicit date
ranges at least once to make everything is up-to-date.

But since then, I added a whole bunch of logs, so we can
see how many user areas are being accessed and filtered each night
(about 10000, then 3000 are skipped because they had no associated
geostore id), but then the number actually processed are capped to the
first 1500 (see end of get_gestore_ids).

So, with Justin, we decided we should fetch the areas using a sorting
option that returns areas with the newest creation time first. See the
change to the url in the page loop of get_pending_areas(). This will
hopefully make sure that we actually get to the most recently created
areas, instead of always looking at old areas, which somehow are not
getting their status changed to 'saved'.

Also fixed some spacing issues in sync.py, do reduce flake8 complaints.