GTC-1250 Update lambda layer based on any datapump source changes
The datapump was using the lambda-layer module of the gfw-lambda-layers
repo. This module only updates a lambda layer if the associated
Dockerfile changes, since it is mainly targeted to creating layers of
existing python libraries, so only the version number in the Dockerfile
matters.

So, changed the terraform to build the layer.zip using the usual
Dockerfile, but using the hash of the entire layer.zip to decide if the
layer.zip should be uploaded to S3 and used to update the lambda. We no
longer use the lambda-layer module.

The layer.zip file includes all the modified times of the files, which
keep changing, even when the file names and contents are the same. I
tried generating a hash using only filenames and contents, but terraform
seems to create its own hash of the layer.zip file as well, so basically
we're always going to update the layer.zip no matter what, which seems
fine.

Added a bunch of explanatory comments, cleaned up some other comments.