GTC-3357 Dist alert datapump changes for int/dist merge
The main thing is adding a job that runs in parallel with the COG jobs
to resample dist alerts raster to 10m resolution. This required making
job.cog_asset_parameters be job.cog_or_aux_asset_parameters that can
handle both COG and aux asset requests.

I also added a new aux_asset job to create a date_conf raster for dist
alerts that has the right date basis so that OTF queries via the data
API can correctly use the __date and __confidence fields. (OTF analysis
on dist was requested by Anika and Sarah.)

I also added a bunch of comments on the sequence of steps for
RasterVersionUpdateJob jobs.