GTC-3375 Datapump changes to create nightly gfw_integrated_dist_alerts (#197)
* GTC-3375 Datapump changes to create nightly gfw_integrated_dist_alerts

Each version of gfw_integrated_dist_alerts is based on the same day's
version of gfw_integrated_dist_alerts and the most recent
umd_glad_dist_alerts (created on Saturday). See the comments for an
explanation of the calc formula. After creating the merged raster, we
then create the intensity raster and the default/intensity COGs.

* Add gfw_integrated_dist_alerts as a SyncType/SYNCER

* Fix up comments based on Daniel's suggestions.