Move back start of nightly runs by 2 hours. (#202)
This is to make it easier to run Pro's weekly runs (which should include
the int-dist raster with the latest dist alerts) on late
Sunday night/early Mon morning.

Also, fixed all the references to EST (eastern standard time) to PST,
which is what the times have always been anyways.