Delete assets of dist alerts versions that are reproducible. (#206)
* Delete assets of dist alerts versions that are reproducible.

We are saving all old weekly dist alert rasters, but delete all the other
assets in the next-to-last version that can be reproduced from the default
raster, including the intensity raster, date_conf raster, COGs, and asset
resampled to 10 meters, etc.

* Don't delete assets of a dist version if it is still latest.

* Remove assets from second-to-last version

Works better if there was a failure in creating the last version, so
next-to-last version is still atest. Failure to create a new complete
dist version are very rare.